Only support "most-recent" context bytes in serial stream.
Removal of an API footgun, as this parameter really only makes
sense for the `serial history` query command, not the continuous
websocket connection.

Eventually over an instance's lifespan, it makes less sense for a
client connecting to try asking for bytes starting from the
beginning of time, as we don't buffer everything forever and this
eventually becomes impossible to satisfy.