Don't require compiling openssl to use the SDK
Currently, crates depending on the oxide SDK cannot choose for
themselves whether to use native TLS or rustls. They also cannot choose
whether to use the system's OpenSSL or use the vendored TLS. This is
because the `native-tls-vendored` feature is specified on reqwest at the
project-level.

This feature is wanted by the cli, which ships binaries bundling openssl
in it. So let's move it so the cli crate turns it on, while the library
crate has no opinion.