Work around clap's default value formatting for instance serial console escape string
Seen in #573.

`clap` is somewhat stubborn about this (intentionally so, with good reason for 99%+ of cases)

- even if we made a custom type for it, values must be able to round-trip
  through being a plain string, one way or another:
  https://docs.rs/clap/latest/clap/_derive/index.html#arg-attributes

- clap unconditionally outputs them in their `OsStr` form,
  which in our case just writes character values that the terminal skips,
  making it appear as an empty string:
  https://github.com/clap-rs/clap/blob/5be548dd54d0a001b1c5d2ad4b10019a062104a4/clap_builder/src/output/help_template.rs#L805
  (note: `to_string_lossy` isn't the culprit, they're valid UTF-8)

The somewhat unusual method we have of passing this parameter when it
contains unprintable characters is motivated in part by wanting to ensure
that the user is actually capable of inputting the character(s) they want
to use, with their keyboard, in the terminal they're using.

However, we do have an escape-hatch (pun unintended):
https://docs.rs/clap/latest/clap/struct.Arg.html#method.hide_default_value

So we'll just hide their version of it and substitute our own.