Print error source on failure in CLI
Currently the `oxide` binary will print the error returned when a
subcommand fails, but not any `source` errors attached to it. In the
context of connection failures from the `auth login` subcommand, the
output was only `Request failed`, giving the user no context on why the
failure occurred.

If present, append the source of the error to the user-facing message.

Previously:
```
$ oxide auth login --host oxide.invalid
Request failed
```

Now:
```
$ oxide auth login --host oxide.invalid
Request failed: error sending request for url \
 (http://oxide.invalid/device/auth): error trying to connect: dns \
 error: failed to lookup address information: nodename nor servname \
 provided, or not known
```