re-organize CLI config files (#727)
The new CLI config files are kept in ~/.config/oxide as credentials.toml (the credentials) and config.toml (other configuration, in particular the default profile). Credentials are organized as "profiles" which may be specified with the --profile option to the CLI. In addition, authentication for the SDK gets much simpler: oxide::Client::new_authenticated() which uses the same configuration files and auth logic as the CLI (or rather, the CLI uses the oxide::Client::new_authenticated_confg() interface which adds customization options).