Add progress spinner to auth status subcommand
When an authenticated Oxide rack is unreachable, e.g. dogfood when the
VPN is disabled, `oxide auth status` may take tens of seconds to
complete. It is unclear to the user which rack is the source of the
delay.

Add a progress spinner that shows the rack currently being checked.

Closes https://github.com/oxidecomputer/oxide.rs/issues/567