Accept profile argument in all subcommands (#781)
The `profile` argument is associated with the top-level `oxide` command,
meaning it must be applied before a subcommand is registered. It would
be convenient to be able to specify this arg at any point in the
command, such as when repeating the same command against different
profiles.

Set `global` to `true` for `profile` to allow this behavior.