Don't recommend saving shell completions to disk
Currently our docs advise users of several shells to save the output of
`oxide completion` to a file and source that in their init. This has led
to users having outdated completions when they upgrade to a new version
of the CLI.

Update the docs to recommend generating the completions on the fly at
startup. Doing this will add a bit of latency to their shell startup,
but user instructions should prioritize reliability over performance.