Indicate if arg is global in docs
With f60ae65 (Accept profile argument in all subcommands (#781),
2024-08-08) we set `--profile` as a global arg, allowing all subcommands
to accept it. This is good for useability, but a bit of a mess in the
docs as it is now listed as an argument for all subcommands.

Removing it entirely from the docs hurts discoverability, but we want to
ensure it is clearly distinguished from the direct args for a
subcommand. Adding a subsection of the docs page with global args would
accomplish this.

Add a new `global` bool for `JsonArg` to indicate whether a given arg is
inherited.