Handle empty credentials.toml files
After logging out of all profiles we write an empty `credentials.toml`
file to the user's config directory. However, when loading that file we
will fail to parse the contents as a TOML table.

Only parse `credentials.toml` if it is not empty, falling back on the
default value otherwise.

Closes https://github.com/oxidecomputer/oxide.rs/issues/797