Add rename-profile subcommand to auth
Currently we name the profile associated with a silo based on the silo
URL. Users who want a more descriptive name must manually edit the
`credentials.toml` and `config.toml` files to update the profile name.
These files are thinly documented and subject to change, a programmatic
way to perform this change would be helpful, particularly for
non-technical users.

Add a new `rename-profile` subcommand to `auth` to update the profile
name and default-profile (when needed).