allow initializing port settings with a link
When a user tries to make a change to a port settings object that does
not exist, we tell them to create a link to initialize the port.
However, when they do just that, we tell them to create a link to
initialize the port. A bit of a jerk move. This fix allows port settings
to be initialized via link creation.