Expand permitted ByteCount formatting (#861)
Currently we require size suffixes to `ByteCount` arguments to be
lowercase and directly after the numerals. Users may expect to pass an
argument with upper case suffixes, e.g., GiB, or with a space before
the unit, e.g., `oxide instance create --memory '1 gib'.

Update the `TypedValueParser` for `ByteCount` to ignore case and permit
spaces within the argument.