Add unit help to all ByteCount fields
The help strings for API fields accepting a `ByteCount` have a dilemma.
For users directly consuming the API, or via the SDK, we need to make it
clear that they should use bytes as the type name suggests. For CLI,
however, we want to encourage users to pass unit suffixes for ease of
use. Help text that emphasizes bytes will lead users to write out values
in bytes, an annoying and error-prone process.

With https://github.com/oxidecomputer/omicron/pull/6737 we have updated
the doc strings on all external API `ByteCount` fields to include the
string "(in bytes)". Add a function to update all CLI `ByteCount` args
and remove that string, appending a help message explaining the units we
accept. While we're at it, also harmonize all CLI-specific `ByteCount`
doc strings.