Set 0600 permissions on credentials.toml
Currently we create the `credentials.toml` file with 0644 permission on
Unix. Given that this file contains sensitive data, 0600 is a more
appropriate permission.

On Windows my rather vague understanding is that the `%USERNAME%`
directory we default to is only accessible by that user (and admins).
There are additional settings like ACLs we could tweak, but these are
not accessible from the Rust standard library and probably more trouble
than they're worth.

We deliberately do not set `config.toml` as 0600 as this does not
contain any sensitive information.