Set mode when creating file
Using `File::create` will create a new file with 0644 permissions, which
potentially leaves an interval where another user could access the file
before we updated it to 0600.

Switch to using `OpenOptions` so we can directly create the file with
the correct permissions. However, in cases where the file already
exists, `OpenOptions::mode` will not overwrite the mode of an existing
file, so we retain the subsequent `fstat` and `fchmod` calls to handle
that scenario.