Add tests for SDK extras and fix `execute`
We've been relying on the CLI disk import tests to cover the `extras`
disk importer in the SDK. However, the CLI does not use the `execute`
method, and we missed that this was failing immediately due to the
cancellation sender being dropped before the request could awaited.

Add tests for `extras` to validate that both options work. Also, add a
method to DiskImport that does not accept cancellation and use that in
`execute`.