Use crates.io over git for dependencies
The presence of git dependencies can be problematic for build systems
like Nix. `progenitor` as a git dependencies and `progenitor-client` as
a registry causes the build to fail because it tries to vendor
`progenitor-client` from two places: from crates.io as a direct
dependency, and from GitHub as a transitive dependency of `progenitor`.

This commit changes the Cargo.toml and lock file to pull `progenitor`
from the registry instead of GitHub.

Fixes #950