Don't treat failure to send as an error
If the file reader task fails to send a chunk to an upload task, we know
this is because that task has either returned due to an error or
panicked. In either case this isn't an error for the reader, which
should return early as the overall request will fail.