disk import: Prioritize returning upload errors (#992)
When sending chunks of the file to upload tasks, we return an error if
the channel to that task is closed. This situation will happen if the
upload task has panicked or already returned due to an error. However,
we check for an error from the sending task first, and the `failed to
send chunks` error will be surfaced to the user, rather than the true
problem that occurred in the upload task.

In the inverse situation, where the reader fails, the uploaders will
return gracefully when their channels are closed, so we do not need to
worry about a problem reading the file showing up as an upload error.

Update our error handling to check for an error from the upload tasks,
then the reader if those returned cleanly.