Bump stack size on Windows
With e31ef2f (Default to current_thread tokio runtime, 2025-02-05) we
converted the CLI to use the `current_thread` tokio executor. We had
seen stack overflows on Windows in the past, which defaults to a 1 MiB
stack. While this problem is not currently present, a PR that increases
the depth of our call stack is encountering it in CI[0].

Add linker flags to bump the stack size to 8 MiB for both MSVC and
MinGW, matching the default value of Linux and macOS.

https://github.com/oxidecomputer/oxide.rs/actions/runs/13817821578/job/38655523269?pr=1053