Set --debug to show only oxide logs
The common case for users using `--debug` will be to observe the logs
generated by the SDK, not the events from hyper and reqwest. If all logs
are needed, we can advise them to use `RUST_LOG=debug` to enable debug
logs for all components.

Set `--debug` to set the log env filter to `oxide=debug`.