Currently only the `auth` CLI subcommand will log any request details when debug logging is enabled. Adding debug logging for all requests will enable users to troubleshoot issues on any API endpoint.
To do this, we use the newly added exec hook feature in Progenitor to
capture the details of the API call.

In addition to the log event, we also create a `Span`. This is ignored
by the CLI, but may be useful for other consumers of the SDK that want
to integrate it with their monitoring infrastructure.

An example event:

  {
    "timestamp": "2025-05-07T19:51:02.464942Z",
    "level": "DEBUG",
    "message": "request succeeded",
    "url": "https://oxide.sys.r3.oxide-preview.com/v1/instances?project=will",
    "path": "/v1/instances",
    "operation_id": "instance_list",
    "remote_addr": "45.154.216.35:443",
    "http.request.method": "GET",
    "http.response.content_length": 2377,
    "http.response.status_code": 200,
    "start_time": "2025-05-07T19:51:02.079228Z",
    "duration_ms": 385,
    "oxide.request_id": "1ff7d872-3a14-4604-8392-1e9d6e2ba8fd",
    "target": "oxide"
  }

Closes #1014