Set longer timeout and read_timeout for bundles (#1107)
Commands that interact with support bundles may be much slower than the
typical external API endpoint, as bundles may reach tens of gigabytes.

If performing a bundle-related subcommand, set a default timeout of 60
minutes. Previously we would set `connect_timeout` and `timeout` to the
same value, but one hour is not reasonable for the initial
`connect_timeout`, so we add a separate config item for that.

To avoid keeping an idle connection open, also add a `read_timeout`
option, which will terminate a connection prior to its timeout when no
data is received.

Only `bundle inspect` and `bundle download` actually interact with the
bundle directly. The other bundle-related subcommands are standard API
calls and can use the normal timeout.

Limit our longer timeouts to just the potentially slow subcommands.

Closes #1095