Take paths instead of base64 for SAML creation
Currently the CLI's SAML IdP creation flow does not provide flags for
signing keys, requiring the cryptic `--json-body` argument. These and
the XML metadata must be manually base64-encoded by the user as well.

Make this process more convenient by converting the keys and XML
metadata args to take the path of the source files, which we then read
and encode ourselves.

While we're at it, hide the no longer needed `--json-body` and
`--json-body-template` arguments in the CLI and stop generating docs
for hidden items.

Example command:

```
$ oxide silo idp saml create \
--silo cli \
--name "keycloak" \
--description "Keybloke" \
--sp-client-id cli \
--acs-url "https://cli.sys.example.com/login/cli/saml/keycloak" \
--idp-entity-id "https://keycloak.example.com/realms/cli" \
--slo-url "" \
--technical-contact-email "user@example.com" \
--group-attribute-name group
--metadata-path ./cli-metadata.xml \
--public-cert-path ./cli-cert.der \
--private-key-path ./cli-key.der \
```