xtask generate: be a bit more helpful about required nightly version (#1124)
while the README says that cargo xtask generate requires a nightly, we actually require that the installed nightly be from a rust-lang/rust commit more recent than the toolchain listed in rust-toolchain.toml.

this means that, say, if you last installed a nightly around Rust 1.82.0, the toolchain upgrade in ee43522 means the next time you cargo xtask generate it would complain that about using unstable rustfmt features. while a nightly rustfmt is installed, even!

with this change, it'll instead end up erroring like "... requires a nightly rustfmt more recent than <date> to be installed", which is hopefully a more direct hint for engineers to unstick themselves.