Add option for tabular output
Currently all commands will display their results as pretty-printed
JSON. Tabular output is convenient for easy comparison of a single field
when multiple objects are returned, and more amenable to traditional
shell processing.

Add a new global `--format` flag to control output format, defaulting to
the existing JSON output. As larger output items, such as instances, can
easily overflow a typical terminal width, we allow users to specify
which fields to print with `--format=table:field1,field2,...`.

Non-scalar fields within a returned object will be printed in compact
JSON format, e.g. `{"cpus":0,"memory":0,"storage":0}` for the
`allocated` field on `oxide silo utilization list`.

To determine the field names to be shown in the table header, we parse
the schema for the return type as part of `OxideOverride`. This logic
makes some assumptions about the shape of the data returned, and we need
to ensure that it remains valid. Add a new return_types `xtask` job,
which writes out all return types from the Oxide API to a file, against
which we test the parsing logic.