Use OXIDE_PROFILE env var to select profile
It can be convenient to select the silo to target via the environment.
We have the `OXIDE_HOST` and `OXIDE_TOKEN` env vars today, but these
require putting your private token in the environment.

Add a new `OXIDE_PROFILE` env var which accepts just the profile name
from `credentials.toml`. To avoid confusion around precedence between
this and `OXIDE_HOST`, return an error if both are set.

Closes https://github.com/oxidecomputer/oxide.rs/issues/1080