Enable version-specific Homebrew formulas (#1240)
The Oxide control plane is in the process of adding backwards
compatibility for clients on older versions. It will be better for a
client to remain on an older version than be ahead of the server.

Homebrew generally uses a "live at head" approach to versioning, which
creates a risk that users will inadvertently update themselves to a
version incompatible with their rack(s). To mitigate this, enable the
newly added `version-formulas` feature in our `dist` fork, which will
create a new formula in our tap for each minor version released.

Users can have multiple versions installed simultaneously, though only
one can be linked to `$HOMEBREW_PREFIX/bin`, the others can be accessed
via `$HOMEBREW_PREFIX/opt/oxide-cli@<VERSION>/bin`.