Read certificate files directly (#1267)
Customers have found it inconvenient and error-prone to pass in the
contents of certs and keys to the `certificate create` subcommand. To
make this command easier to use, update its `key` and `cert` arguments
to be treated as the path to their respective files, rather than their
contents.

We previously updated the SAML IdP creation command the same way with
11fe44b (Take paths instead of base64 for SAML creation (#1112),
2025-05-28).

Co-authored-by: Adam Leventhal <ahl@oxide.computer>